Create Ui.java for OOP
