+++ src/main/java/piper/Piper.java
@@ -1,0 +1,23 @@
+package piper;
+
+import piper.ui.Ui;
+
+public class Piper {
+    public static void main(String[] args) {
+        final String CHATBOT_NAME = "Piper";
+        Ui ui = new Ui(CHATBOT_NAME);
+        boolean exit = false;
+
+        ui.greetUser();
+
+        while (!exit) { // user is active
+            String userInput = ui.read();
+            if (userInput.equals("bye")) { // user is inactive
+                ui.farewellUser();
+                exit = true;
+            } else {
+                ui.echoUser(userInput);
+            }
+        }
+    }
+}
